change Build.scala to build.sbt
